fade tests -> doctest #202
